requirements.txt
# Existing requirements
osmnx>=1.3.0
networkx>=3.1
matplotlib>=3.7.1
folium>=0.14.0
requests>=2.31.0
PyYAML>=6.0.1
pytest>=7.3.1
geopy>=2.3.0

# New web dependencies
Flask>=2.0.1
Flask-Cors>=3.0.10
gunicorn>=20.1.0  # For production deployment